[Ops] Fix int32 overflow in pointer arithmetic across all Triton kernels
Cast batch/sequence indices to int64 before stride multiplication in ~83
Triton kernel files. tl.program_id() returns int32; when multiplied by
large strides the intermediate products overflow INT32_MAX, causing
illegal CUDA memory accesses or silently wrong results.

Also adds AGENTS.md with review guidelines to help agent code reviewers
keep future Triton kernels scale-friendly.

+++ fla/ops/cp/chunk_delta_h.py
@@ -327,9 +327,9 @@ def merge_fwd_bwd_kernel(
             return
 
         # Load offsets for this sequence
-        ss_start = tl.load(seq_offsets + i_seq).to(tl.int32)
-        ss_end = tl.load(seq_offsets + i_seq + 1).to(tl.int32)
-        init_base = tl.load(init_offsets + i_seq).to(tl.int32)
+        ss_start = tl.load(seq_offsets + i_seq).to(tl.int64)
+        ss_end = tl.load(seq_offsets + i_seq + 1).to(tl.int64)
+        init_base = tl.load(init_offsets + i_seq).to(tl.int64)
         num_subseqs = ss_end - ss_start
 
         stride_hm_s = HV * K * (V + K)
