+++ fla/ops/generalized_delta_rule/dplr/chunk_A_bwd.py
@@ -67,10 +67,10 @@ def chunk_dplr_bwd_kernel_intra(
     i_b, i_h = i_bh // H, i_bh % H
     if IS_VARLEN:
         i_n, i_t = tl.load(chunk_indices + i_t * 2).to(tl.int32), tl.load(chunk_indices + i_t * 2 + 1).to(tl.int32)
-        bos, eos = tl.load(cu_seqlens + i_n).to(tl.int32), tl.load(cu_seqlens + i_n + 1).to(tl.int32)
-        T = eos - bos
+        bos, eos = tl.load(cu_seqlens + i_n).to(tl.int64), tl.load(cu_seqlens + i_n + 1).to(tl.int64)
+        T = (eos - bos).to(tl.int32)
     else:
-        bos, eos = (i_b * T).to(tl.int32), (i_b * T + T).to(tl.int32)
+        bos, eos = (i_b * T).to(tl.int64), (i_b * T + T).to(tl.int64)
 
     if i_t * BT >= T:
         return
@@ -273,10 +273,10 @@ def chunk_dplr_bwd_kernel_intra_tensorcore(
 
     if IS_VARLEN:
         i_n, i_t = tl.load(chunk_indices + i_t * 2).to(tl.int32), tl.load(chunk_indices + i_t * 2 + 1).to(tl.int32)
-        bos, eos = tl.load(cu_seqlens + i_n).to(tl.int32), tl.load(cu_seqlens + i_n + 1).to(tl.int32)
+        bos, eos = tl.load(cu_seqlens + i_n).to(tl.int64), tl.load(cu_seqlens + i_n + 1).to(tl.int64)
         T_len = eos - bos
     else:
-        bos, eos = (i_b * T).to(tl.int32), (i_b * T + T).to(tl.int32)
+        bos, eos = (i_b * T).to(tl.int64), (i_b * T + T).to(tl.int64)
         T_len = T
 
     if i_t * BT >= T_len:
@@ -424,13 +424,13 @@ def chunk_dplr_bwd_dgk_kernel(
     if IS_VARLEN:
         i_tg = i_t
         i_n, i_t = tl.load(chunk_indices + i_t * 2).to(tl.int32), tl.load(chunk_indices + i_t * 2 + 1).to(tl.int32)
-        bos, eos = tl.load(cu_seqlens + i_n).to(tl.int32), tl.load(cu_seqlens + i_n + 1).to(tl.int32)
-        T = eos - bos
+        bos, eos = tl.load(cu_seqlens + i_n).to(tl.int64), tl.load(cu_seqlens + i_n + 1).to(tl.int64)
+        T = (eos - bos).to(tl.int32)
         NT = tl.cdiv(T, BT)
     else:
         NT = tl.cdiv(T, BT)
         i_tg = (i_b * NT + i_t).to(tl.int32)
-        bos, eos = (i_b * T).to(tl.int32), (i_b * T + T).to(tl.int32)
+        bos, eos = (i_b * T).to(tl.int64), (i_b * T + T).to(tl.int64)
 
     stride_qk = H * K
     dgk += (bos * H + i_h) * K
